Add type annotations
These are not quite as tight as I would like them to be: for example,
returning Any from json.loads() permits a much more flexible return
object when the code actually has some expectations about its
structure.

Note that in the gmail code i've also changed the arguments to
log_request, because it was missing a self argument.  The function
ignored all aguments, of course, so it was probably OK, but better to
match the baseclass definition.